requirements.txt
fastapi 
uvicorn[standard] 
cryptography 
argon2-cffi 
